* moved ccd types out into its own file * added some more intelligent type defs * removed ts-ignores * still waiting on a type def for Answers
