+++ journeys/et/configRead.ts
@@ -1,4 +1,4 @@
-import { Journey } from "types/types"
+import { Journey } from "types/journey"
 import { getReadTime, readInCurrentConfig } from "app/et/configs"
 import { formatMsSpanNearestUnit } from "@spacepumpkin/format-timespan"
 import { restorePreviousSession, session, SESSION_EXT } from "app/session"
