+++ journeys/et/createEventToComplexType.ts
@@ -1,5 +1,5 @@
 import { prompt } from "inquirer"
-import { CaseEventToFieldKeys, CaseFieldKeys, EventToComplexTypeKeys, Journey } from "types/types"
+import { CaseEventToFieldKeys, CaseFieldKeys, EventToComplexTypeKeys } from "types/ccd"
 import { createSingleField, askCaseEvent } from "./createSingleField"
 import { createNewEventToComplexType } from "app/objects"
 import { addToInMemoryConfig, getKnownCaseFieldIDs } from "app/et/configs"
@@ -7,6 +7,7 @@ import { fuzzySearch } from "app/questions"
 import { CUSTOM } from "app/constants"
 import { session } from "app/session"
 import { getIdealSizeForInquirer } from "app/helpers"
+import { Journey } from "types/journey"
 
 const QUESTION_CASE_EVENT_ID = "What event does this belong to?"
 const QUESTION_ID = "What's the ID of this EventToComplexType?"
