+++ journeys/et/dockerFreshStart.ts
@@ -1,5 +1,5 @@
 import { ensureUp, destroyEverything } from "app/et/docker"
-import { Journey } from "types/types"
+import { Journey } from "types/journey"
 
 async function tearDownAndFreshStart() {
   await destroyEverything()
