+++ journeys/et/sessionSetName.ts
@@ -1,6 +1,6 @@
 import { prompt } from "inquirer"
 import { session } from "app/session"
-import { Journey } from "types/types"
+import { Journey } from "types/journey"
 
 const QUESTION_NAME = "What should we called this session?"
